added typing file for auth context and changed auth context wrapper
