+++ client/src/types/AuthContext.d.ts
@@ -1,0 +1,12 @@
+export type AuthContextType = {
+  user: User;
+  setUser: (user: User) => void;
+};
+
+type User = {
+  id: number;
+  first_name: string;
+  last_name: string;
+  email: string;
+  role: "parent" | "nursery";
+} | null;
